mouvement ennemi fini (mais pas visible)
il faut utiliser ton imagination pour les voir

+++ src/graphe.hpp
@@ -15,8 +15,8 @@ namespace Graph {
 
     struct Node {
         int id {0};
-        int x {0};
-        int y {0};
+        float x {0};
+        float y {0};
         std::vector<WeightedGraphEdge> edges;
     };
